Rename lra project to io.openliberty
