+++ dev/com.ibm.websphere.appserver.features/visibility/public/mpLRA-1.0/com.ibm.websphere.appserver.lra-1.0.feature
@@ -12,7 +12,7 @@ IBM-API-Package: \
 -features=com.ibm.websphere.appserver.org.eclipse.microprofile.lra-1.0, \
           com.ibm.websphere.appserver.jaxrs-2.1
 -bundles=com.ibm.ws.require.java8, \
-         com.ibm.ws.lra, \
+         io.openliberty.mpLRA, \
          com.ibm.ws.org.jboss.narayana.rts
 kind=noship
 edition=full
