+++ dev/io.openliberty.mpLRA.1.0/.settings/bndtools.core.prefs
