+++ dev/io.openliberty.mpLRA.1.0/resources/OSGI-INF/l10n/metatype.properties
