+++ dev/io.openliberty.mpLRA.1.0/resources/OSGI-INF/metatype/metatype.xml
